Refactored the components for watching a directory into a single class.
